refactor: centralizar utilidades de transpiladores
