+++ src/cobra/transpilers/__init__.py
@@ -1,10 +1,6 @@
 """Facilita el acceso a los distintos transpiladores de Cobra."""
 
-from cobra.transpilers.base import BaseTranspiler
+from cobra.transpilers.common.utils import BaseTranspiler
 from cobra.transpilers.reverse.base import BaseReverseTranspiler
 
 __all__ = ["BaseTranspiler", "BaseReverseTranspiler"]
-
-
-def exceptions():
-    return None
