+++ src/cobra/transpilers/base.py
@@ -1,24 +1,8 @@
-from abc import ABC, abstractmethod
-from core.visitor import NodeVisitor
+"""Compatibilidad retrocompatible para :mod:`cobra.transpilers`.
 
+El contenido real se encuentra en ``cobra.transpilers.common.utils``.
+"""
 
-class BaseTranspiler(NodeVisitor, ABC):
-    """Clase base para los transpiladores de Cobra."""
+from cobra.transpilers.common.utils import BaseTranspiler, save_file
 
-    def __init__(self):
-        self.codigo = []
-
-    @abstractmethod
-    def generate_code(self, ast):
-        """Genera el código a partir del AST proporcionado."""
-        raise NotImplementedError
-
-    def save_file(self, path):
-        """Guarda el código generado en la ruta dada."""
-        contenido = (
-            "\n".join(self.codigo)
-            if isinstance(self.codigo, list)
-            else str(self.codigo)
-        )
-        with open(path, "w", encoding="utf-8") as archivo:
-            archivo.write(contenido)
+__all__ = ["BaseTranspiler", "save_file"]
