+++ src/cobra/transpilers/common/__init__.py
@@ -1,0 +1,17 @@
+"""Módulo con utilidades compartidas por los transpiladores."""
+
+from .utils import (
+    BaseTranspiler,
+    get_standard_imports,
+    load_mapped_module,
+    save_file,
+    STANDARD_IMPORTS,
+)
+
+__all__ = [
+    "BaseTranspiler",
+    "get_standard_imports",
+    "load_mapped_module",
+    "save_file",
+    "STANDARD_IMPORTS",
+]
