+++ src/cobra/transpilers/transpiler/to_cpp.py
@@ -27,7 +27,7 @@ from cobra.core.ast_nodes import (
 )
 from cobra.core import TipoToken
 from core.visitor import NodeVisitor
-from cobra.transpilers.base import BaseTranspiler
+from cobra.transpilers.common.utils import BaseTranspiler
 from core.optimizations import optimize_constants, remove_dead_code, inline_functions
 from cobra.macro import expandir_macros
