+++ src/cobra/transpilers/transpiler/to_java.py
@@ -14,7 +14,7 @@ from cobra.core.ast_nodes import (
     NodoInstancia,
 )
 from cobra.core import TipoToken
-from cobra.transpilers.base import BaseTranspiler
+from cobra.transpilers.common.utils import BaseTranspiler
 from core.optimizations import optimize_constants, remove_dead_code, inline_functions
 from cobra.macro import expandir_macros
