+++ src/cobra/transpilers/transpiler/to_latex.py
@@ -10,7 +10,7 @@ from cobra.core.ast_nodes import (
     NodoAtributo,
 )
 from core.visitor import NodeVisitor
-from cobra.transpilers.base import BaseTranspiler
+from cobra.transpilers.common.utils import BaseTranspiler
 from core.optimizations import optimize_constants, remove_dead_code, inline_functions
 from cobra.macro import expandir_macros
